+++ packages/babel-preset/CHANGELOG.md
@@ -9,7 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Breaking Change
 
-- Removed `web`, `node`, and `react` presets for `common`. See [documentation](./README.md) for usage. [[#252](https://github.com/Shopify/web-configs/pull/252)]
+- Removed `web`, `node`, `react`, and `common` presets for a single `@shopify/babel-preset` import. See [documentation](./README.md) for usage. [[#252](https://github.com/Shopify/web-configs/pull/252)]
 
 ## 23.6.1 - 2021-05-05